+++ documentation/code/constants.ts
@@ -1,7 +1,6 @@
 export const OUTER_CONTAINER_STYLE = {
   background: 'white' as 'white',
   padding: '20px',
-  boxSizing: 'border-box' as 'border-box',
-  height: '400px',
-  overflow: 'hidden' as 'hidden',
+  borderRadius: '6px',
+  border: '2px solid #EAECEF',
 };