+++ src/components/LineChart/components/index.ts
@@ -1,4 +1,3 @@
-export {Legend} from './Legend';
 export {Line} from './Line';
 export {GradientArea} from './GradientArea';
 export {TooltipContent, TooltipContentProps} from './TooltipContent';