+++ src/components/MultiSeriesBarChart/components/index.ts
@@ -1,3 +1,2 @@
 export {BarGroup} from './BarGroup';
-export {Legend} from './Legend';
 export {StackedBarGroup} from './StackedBarGroup';